mdb: add `reformat` command
The `reformat` command takes a JSON document and reformats it in a
deterministic way. Right now, it orders dictionary keys and streams
everything through the python JSON implementation.

This will be used in the future by our CI to verify all JSON documents
have a standardized format (if needed). Since we often rely on checksums
for documents, this way we avoid adding semnatical duplicates to the
repository, or accidentally failing an assertion just because we got
non-deterministic JSON output.

The `mdb reformat` tool is meant to be used by the CI on all stored
applicable documents to assert they have been committed correctly. It
will not be used at runtime, unless necessary.